+++ src/product-service/handlers/getProductsList.test.ts
@@ -1,10 +1,23 @@
-import { handler as getProductsList} from './getProductsList'
-import { productsLambdaTest } from "../../product-service/mocks/data";
-describe('getProductsList', () => {
+import { handler } from "./getProductsList";
+import { getProductsList } from "../../product-service/pg-db/products";
 
-  it('getProductsList test', async () => {
-    const data = await getProductsList();
+jest.mock("../../product-service/pg-db/products");
+
+const ProductsMock = [
+  {
+    count: 3,
+    description: "desc 1",
+    id: "2",
+    price: 20,
+    title: "Title1",
+  },
+];
+describe("getProductsList", () => {
+  it("getProductsList test", async () => {
+    getProductsList.mockReturnValueOnce(Promise.resolve(ProductsMock));
+    const data = await handler();
     expect(data.statusCode).toBe(200);
-    expect(JSON.parse(data.body).length).toBe(productsLambdaTest.length);
+    expect(data.body).toStrictEqual(JSON.stringify(ProductsMock));
+    expect(JSON.parse(data.body).length).toBe(1);
   });
 });
